requirements.txt
requests==2.32.3
Jinja2==3.1.4
Flask==3.0.3
gunicorn==23.0.0
python-dotenv==1.0.1